nbx is uninitialized in certain code paths
The nbx local variable is only initialized in
jtjprx if a box character is on the parse line.

As a result the conditional that checks if a
fast path copy can be done will fall back to
the slow path most times due to the unlikely
event if nbx being randomly initialized to zero.

To test, run under valgrind and enter at the
console:

    1

This will trigger the unitialized variable warning.
The following does not trigger it as it contains
the box character:

    <1

Fix is to initialize nbx to zero on entry to function.