Allow jtjfread to work on pipes
jtjfread can attempt to do an fseek on the passed FILE pointer via
a call to fsize. fsize doesn't do any error checking of the fseek
call on 64 bit builds and returns an uninitialized value in that case.

fseek will fail if its passed a FILE pointer to a pipe endpoint or
other non-seekable stream. The hostio foreign conjunction, 2!:2,
returns pipe FILE pointers. This prevents 1!:1 from working with
these pointers.

The workaround here is to do an ftell on the file, which will fail
in the case of non-seekable streams. If it fails then rdns is called
instead of rd. rdns does not use fseek and instead reads all contents
of the stream.

Error checking is also added to fsize to return a consistent value of
-1 on error, following the same behaviour as the non-64 bit case.